refactor(web): extract pure turn-rendering helpers from ChatPane
